+++ src/adapters/createMatchingRouterState.ts
@@ -15,21 +15,25 @@ export const createMatchingRouterState = (
     queryParseOptions?: any
 ): RouterState | undefined => {
     // Find the matching route
-    let matchingRoute = null;
-    let params = undefined;
-    for (let i = 0; i < routes.length; i++) {
-        const route = routes[i];
-        params = matchUrl(location.pathname, route.pattern);
-        if (params) {
-            matchingRoute = route;
-            break;
+    for (let rIndex = 0; rIndex < routes.length; rIndex++) {
+        const route = routes[rIndex];
+
+        // Combine primary and alternate patterns into a single array.
+        // Primary should always be at the beginning of the array.
+        const patterns = route.altPatterns
+            ? [route.pattern].concat(route.altPatterns)
+            : [route.pattern];
+
+        // Match URL to the patterns
+        for (let pIndex = 0; pIndex < patterns.length; pIndex++) {
+            const params = matchUrl(location.pathname, patterns[pIndex]);
+            if (params) {
+                const queryParams = parse(location.search, queryParseOptions);
+                return createRouterState(route.name, {
+                    params,
+                    queryParams,
+                });
+            }
         }
-    }
-    if (matchingRoute) {
-        const queryParams = parse(location.search, queryParseOptions);
-        return createRouterState(matchingRoute.name, {
-            params,
-            queryParams,
-        });
     }
 };